I have client side store of viewstate data now so I dont need to worry about caching
The issue? Maximize breaks sometimes when the tabs are opened in
particular orders and not in others and I am not sure what I did.

However it does not seem to be based on time, so it is not a race
condition. It is a logic error somewhere, possibly to do with pointers.